projects/visitor-estimation/serve/pyproject.toml
[tool.poetry]
name = "visitor-estimation-serve"
version = "0.1.0"
description = "Implementation of visitor estimation API."
authors = ["Zheyuan Hu <zheyuan.hu@onclusive.com>"]
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = "3.9.19"
core = {path = "../../../libs/core"}
models = {path = "../../../libs/models"}
serving = {path = "../../../libs/serving"}
tracking = {path = "../../../libs/tracking"}

[tool.poetry.group.dev.dependencies]
pytest = "7.2.0"
pytest-mock = "3.10.0"
pytest-order = "1.1.0"
requests-toolbelt = "^1.0.0"

[tool.poetry.group.debug.dependencies]
debugpy = "1.8.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[[tool.poetry.source]]
name = "aws-neuron"
url = "https://pip.repos.neuron.amazonaws.com/"
priority = "supplemental"
